Use TPL to process jobs
